Read double quoted strings
I don't know whether it makes sense to check the quoting rules now, or
postpone it.  It boils down to whether I want the lexer to "step into"
the double quoted strings or just return their content for now.  If I
choose the latter, the parser can always instantiate another Lexer and
use that to tokenize the contents, maybe.